% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_sample.R
\name{dag_sample}
\alias{dag_sample}
\title{Generate data from a DAG}
\usage{
dag_sample(DAG, size = 10, seed = NULL)
}
\arguments{
\item{DAG}{list of formulas. Variables must be listed before they
can be used to create other variables.}

\item{size}{size of the sample, that is, the number of rows to be put in the data frame}

\item{seed}{Set the random number seed. Useful for reproducibility.}
}
\description{
A DAG is represented as a list of formulas. The right-hand side
gives the name of the variable. Variables starting with dots will not be
printed.
}
\examples{
dag_sample(dag03)

}
