% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagsim.R
\name{dagsim}
\alias{dagsim}
\title{Generate data from a DAG}
\usage{
dagsim(DAG, nrows = 10, seed = NULL)
}
\arguments{
\item{DAG}{list of formulas. Variables must be listed before they
can be used to create other variables.}

\item{nrows}{Number of rows in the data frame}

\item{seed}{Set the random number seed. Useful for reproducibility.}
}
\description{
A DAG is represented as a list of formulas. The right-hand side
gives the name of the variable. Variables starting with dots will not be
printed.
}
\details{
Some random number generators are built in, and can be used in the DAG.
\itemize{
\item \code{eps(sd=1)} --- normally distributed with the given standard deviation
\item \code{unif(min=0, max=1)} --- uniformly distributed
\item \code{tdist(df=3, ncp=0)} --- t-distributed (long tails)
Also,
\item \code{seq()} --- sequence from 1 to nrows
\item \code{binom(input)} --- generates a 0/1 variable. Probability of 1 is a logistic transformation
of the input.
}
}
\examples{
dagsim(list(x ~ eps(0.5), y ~ x + eps(0.5)))
dagsim(list(.genes ~ eps(), x ~ .genes + eps(), y ~ .genes + eps()))
dagsim(list(x ~ c("a", "b", "c"), y ~ 2), nrow=5) # e.g. for blocking

}
