% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagsim.R
\name{dagsim}
\alias{dagsim}
\alias{sample.dagsystem}
\title{Generate data from a DAG}
\usage{
dagsim(DAG, size = 10, seed = NULL)

\method{sample}{dagsystem}(x, size, replace = FALSE, ...)
}
\arguments{
\item{DAG}{list of formulas. Variables must be listed before they
can be used to create other variables.}

\item{seed}{Set the random number seed. Useful for reproducibility.}

\item{nrow}{Number of rows in the data frame}
}
\description{
A DAG is represented as a list of formulas. The right-hand side
gives the name of the variable. Variables starting with dots will not be
printed.
}
